chore(users): Renaming file and handling missing errors
