+++ ui/src/views/settings/modals/remove-user-modal/index.tsx
@@ -3,7 +3,7 @@ import { Button, Modal, Toolbar } from '@mergestat/blocks'
 import { TrashIcon, XIcon } from '@mergestat/icons'
 import React, { useCallback } from 'react'
 import { RemoveUserMutation } from 'src/api-logic/graphql/generated/schema'
-import { REMOVE_USER } from 'src/api-logic/graphql/mutations/add-users'
+import { REMOVE_USER } from 'src/api-logic/graphql/mutations/manage-users'
 import { useUserSettingsContext, useUserSettingsSetState } from 'src/state/contexts'
 import { showSuccessAlert } from 'src/utils/alerts'
 import { TEST_IDS } from 'src/utils/constants'
